deleted debug_game_state; changed the config_version implementation on the server a little bit; changed the player coordinates type from f64 to i32

+++ client/src/data.rs
@@ -5,8 +5,8 @@ pub mod game_state {
     pub struct PlayerData {
         pub nick: String,
         pub direction: u16,
-        pub x: f64,
-        pub y: f64
+        pub x: i32,
+        pub y: i32
     }
 
     #[derive(serde::Deserialize,Clone,Debug)]
