+++ server/src/game_state.rs
@@ -5,8 +5,8 @@ pub struct PlayerData {
     pub nick: String,
     pub next_move_direction: Option<u8>,
     pub direction: u16,
-    pub x: f64,
-    pub y: f64
+    pub x: i32,
+    pub y: i32
 }
 
 #[derive(Clone,Debug)]
